+++ libs/config/include/mrpt/utils/CConfigFileBase.h
@@ -1,6 +1,0 @@
-#include <mrpt/config/CConfigFileBase.h>
-#pragma message("backwards-compat: update to <mrpt/config/CConfigFileBase.h>")
-namespace mrpt::utils
-{
-using CConfigFileBase = mrpt::config::CConfigFileBase;
-}