+++ libs/io/include/mrpt/utils/CFileGZInputStream.h
@@ -1,6 +1,0 @@
-#include <mrpt/io/CFileGZInputStream.h>
-#pragma message("backwards-compat: update to <mrpt/io/CFileGZInputStream.h>")
-namespace mrpt::utils
-{
-using CFileGZInputStream = mrpt::io::CFileGZInputStream;
-}