+++ libs/system/include/mrpt/utils/CRateTimer.h
@@ -1,6 +1,0 @@
-#include <mrpt/system/CRateTimer.h>
-#pragma message("backwards-compat: update to <mrpt/system/CRateTimer.h>")
-namespace mrpt::utils
-{
-using CRateTimer = mrpt::system::CRateTimer;
-}