+++ doc/source/doxygen-docs/changelog.md
@@ -4,6 +4,7 @@
 - Incompatible API changes:
   - mrpt::system::TParameters has been removed, superseded by mrpt::containers::yaml.
   - Remove mrpt::hwdrivers::CRovio
+  - Removed old mrpt 1.5.x backwards-compatible `<mrpt/utils/...>` headers (Closes #1083).
 - Changes in libraries:
   - \ref mrpt_containers_grp
     - New class mrpt::containers::yaml for nested, YAML-like data structures.